[fix] : 리뷰 평점 평균 구하는 코드
코드 수정

+++ src/main/java/com/kkosoonnae/jpa/projection/StoreDetailViewProjection.java
@@ -16,12 +16,12 @@ public record StoreDetailViewProjection(
         LocalTime openingTime,
         LocalTime closingTime,
         List<String> img,
-        Integer averageScope,
+        Double averageScope,
         Long totalLikeStore
 ) {
     public StoreDetailViewProjection(Integer storeNo, String storeName, String content, String phone,
                                      String roadAddress, LocalTime openingTime, LocalTime closingTime,
-                                     Integer averageScope, Long totalLikeStore) {
+                                     double averageScope, Long totalLikeStore) {
         this(storeNo, storeName, content, phone, roadAddress, openingTime, closingTime,
                 new ArrayList<>(), averageScope, totalLikeStore);
     }
